Working on rational value support
... where values are stored in a nominator/denominator fashion

+++ src/exif_tag_format.rs
@@ -5,12 +5,12 @@ pub type INT8U          = Vec<u8>;
 pub type STRING         = String;
 pub type INT16U         = Vec<u16>;
 pub type INT32U         = Vec<u32>;
-pub type RATIONAL64U    = Vec<u32>;     // ???
+pub type RATIONAL64U    = Vec<f64>;
 pub type INT8S          = Vec<i8>;
 pub type UNDEF          = Vec<u8>;      // got no better idea for this atm
 pub type INT16S         = Vec<i16>;
 pub type INT32S         = Vec<i32>;
-pub type RATIONAL64S    = Vec<i32>;     // ???
+pub type RATIONAL64S    = Vec<f64>;
 pub type FLOAT          = Vec<f32>;
 pub type DOUBLE         = Vec<f64>;
